args , tags , logs
we can run the project with Args
we can use the tags like in_dev, or env, etc.
added allure report generated from Junit file of the result of the run
finished the log file, that create a several files like info file, debug, error, warning
we started to do the environment class to do the before, after step, scenario, feature we will continue work on it, and the rest of the files.